[sonic_package_manager] flush once finished saving docker image into temporary file (#1638)
Since installation happens inside context manager "with tempfile.NamedTemporaryFile()"
it may be that depending on file size/number of writes the docker image isn't fully written
to the disk. In practice a small test docker image was succesfully migrated but when trying with
larger onces it started to file. This change fixes the issue by adding a flush.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>