[config] Remove "reset failed" print lines from config reload (#1654)
#### What I did
I removed the following line from the `config reload` flow....

```
click.echo("Resetting failed status on {}".format(service))
```

This is called in the `_reset_failed_services()` function which is used to reset the systemd restart timers on services so that it doesn't prevent them from starting from the config reload. 

The above line results in the following lines being printed to the terminal when `config reload` is run

```
Resetting failed status on <service>.service
```

This gives the impression that the listed services failed at some point in their lifetime which is untrue and misleading to print to the terminal (in addition many rich-text terminals mark keywords such as "failed" red which further increases the implied severity of these log lines). I do not believe this operation needs any verbosity as it is a very simple operation that makes no user-visible changes thus the user should not need to be notified of it.